Simplify transforming query results into a response
