+++ src/items/todo.rs
@@ -62,7 +62,7 @@ impl Todo {
 mod routes {
     use actix_web::{post, web, Error, HttpResponse};
 
-    use crate::items::Responsable;
+    use crate::utils::responsable::Responsable;
     use crate::{database::exec_on_pool, DbPool};
 
     use super::{NewTodo, Todo};
