+++ src/items/todo_item.rs
@@ -64,7 +64,7 @@ impl TodoItem {
 mod routes {
     use actix_web::{post, web, Error, HttpResponse};
 
-    use crate::items::Responsable;
+    use crate::utils::responsable::Responsable;
     use crate::{database::exec_on_pool, DbPool};
 
     use super::{NewTodoItem, TodoItem};
